Bugfix: Move offset after vbyte write
In the positional index, when a new block is created, vbyte is used to encode first the document gap, and then the word gap. However, the write offset was not updated which would mean that once the document gap was written, it would subsequently be stomped on by the word gap.